Added. Validators for category, event and todo apis.

+++ app/pages/api/[id]/category/delete.ts
@@ -15,22 +15,48 @@ export default async function DeleteCategory(
         /** check user authorisation */
         await authorisedValidator(req, res);
 
-        /** deconstruct body data */
-        const { categoryID } = req.body;
+        try {
+            /** deconstruct body data */
+            const { categoryID } = req.body;
 
-        /* delete data into category table */
-        const result = await executeQuery({
-            query: 'DELETE FROM category WHERE AccountID=? AND ID=?',
-            values: [req.query.id, categoryID],
-        });
+            if (categoryID.length === 36) {
+                /* delete data into category table */
+                const result = await executeQuery({
+                    query: 'DELETE FROM category WHERE AccountID=? AND ID=?',
+                    values: [req.query.id, categoryID],
+                });
 
-        res.status(200).json({ message: 'success' })
-        return
+                res.status(200).json({ message: 'success' })
+                return
+            }
+            /** if request body components do not fit requirements */
+            else {
+                res.statusCode = 400;
+                res.end('Request format error');
+            }
+        }
+        /** if request body components do not fit requirements */
+        catch {
+            res.statusCode = 400;
+            res.end('Request format error');
+        }
     }
-    /* rejects requests that are empty */
-    else if (!req.body) {
+    /* rejects requests that are not POST */
+    else if (req.method !== 'POST') {
         res.statusCode = 405;
         res.end('Error');
+        return
+    }
+    /** if request body components do not fit requirements */
+    else if (!req.body) {
+        res.statusCode = 400;
+        res.end('Request format error');
+        return
+    }
+    /** if user is not authenticated */
+    else if (!req.cookies['token']) {
+        res.statusCode = 401;
+        res.end('Unauthorised');
         return
     }
 }
